Add comments to products and started search

+++ app/helpers/products_helper.rb
@@ -1,10 +1,14 @@
 module ProductsHelper
 
 	def like_link(product)
-		if current_user.liked?(product)
-			link_to "Unlike", unlike_store_product_path(product), method: :delete
-		else
-			link_to "Like", like_store_product_path(product), method: :post
+		if user_signed_in?
+			if current_user.liked?(product)
+				link_to "Unlike", unlike_store_product_path(product), method: :delete
+			else
+				link_to "Like", like_store_product_path(product), method: :post
+			end
 		end
+	else
+		"Sign up to like"
 	end
 end
